Make excuateble
piper is a command line tool that shows you how the processes on your system are connected with pipes! Why? Sometimes you have loggers on your system that communicate through pipes! You might want to know how they all fit together. This is more likely to be useful on a server than on your desktop.